backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9


# Background tasks
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Storage
boto3==1.34.0
minio==7.2.0

# Image and video processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# AI and ML
google-generativeai==0.7.2
openai==1.3.7

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

